core: Reuse unix sockets
